requirements.txt
Django==5.0.6
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-extensions==3.2.3
djangorestframework==3.15.1
prompt_toolkit==3.0.45
psycopg2-binary
pydot==2.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
tzdata==2024.1
python-dotenv